requirements.txt
configparser==7.0.0
praw==7.7.1
pandas==2.2.2
minio==7.2.7
apache-airflow-providers-amazon==8.27.0